Read log level from both user and system config
Resolves issue #1728
