+++ mycroft/util/json_helper.py
@@ -15,6 +15,23 @@
 import json
 
 
+def merge_dict(base, delta):
+    """
+        Recursively merging configuration dictionaries.
+
+        Args:
+            base:  Target for merge
+            delta: Dictionary to merge into base
+    """
+
+    for k, dv in delta.items():
+        bv = base.get(k)
+        if isinstance(dv, dict) and isinstance(bv, dict):
+            merge_dict(bv, dv)
+        else:
+            base[k] = dv
+
+
 def load_commented_json(filename):
     """ Loads an JSON file, ignoring comments
